refactor pulling code
